server/requirements.txt
flask
python-dotenv
psycopg2-binary
flask_cors
flask_login
bcrypt